+++ CHANGE-LOG.md
@@ -58,3 +58,4 @@
 - Remove `useStore` hook which leads to unnecessary renders
 - Action functions are now referentially stable with loading state properties mutated in place on each render.
 - Prettify types and fix many broken compound types by replacing conditional with mapped types
+- Allow `store.setList(null)` to clear list